+++ tests/lib.rs
@@ -1,3 +1,5 @@
+#[macro_use]
+extern crate rulinalg;
 extern crate rusty_machine as rm;
 extern crate num as libnum;
 
@@ -11,3 +13,5 @@ pub mod learning {
     	mod grad_desc;
     }
 }
+
+pub mod datasets;