+++ CHANGELOG.md
@@ -2,6 +2,72 @@
 
 This document will be used to keep track of changes made between release versions. I'll do my best to note any breaking changes!
 
+## 0.5.4
+
+### New Contributors
+
+- [sinhrks](https://github.com/sinhrks)
+
+### Breaking Changes
+
+- None
+
+### Features
+
+- Add a new `datasets` module behind a `datasets` feature flag.
+- Add new classification scores: `precision`, `recall`, and `f1`.
+- Add a new `Transformer::fit` function to allow prefitting of a
+`Transformer` before use.
+
+### Bug Fixes
+
+- None
+
+### Minor Changes
+
+- `LinRegressor` now uses `solve` instead of `inverse` for improved
+accuracy and stability.
+
+## 0.5.3
+
+### Breaking Changes
+
+- None
+
+### Features
+
+- Adding a new `confusion_matrix` module.
+
+### Bug Fixes
+
+- None
+
+### Minor Changes
+
+- Updated rulinalg dependency to `0.3.7`.
+
+## 0.5.2
+
+### New Contributors
+
+- [scholtzan](https://github.com/scholtzan)
+
+### Breaking Changes
+
+- None
+
+### Features
+
+- None
+
+### Bug Fixes
+
+- Regularization constant for GMM is now only added to diagonal.
+
+### Minor Changes
+
+- Added some better `Result` handling to GMM.
+
 ## 0.5.1
 
 This version includes no changes but is a bump due to a
@@ -13,13 +79,13 @@ See the notes for 0.5.0 below.
 
 This is another fairly large release. Thank you to everyone who contributed!
 
-## New Contributors
+### New Contributors
 
 - [NivenT](https://github.com/NivenT)
 - [theotherphil](https://github.com/theotherphil)
 - [andrewcsmith](https://github.com/andrewcsmith)
 
-## Breaking Changes
+### Breaking Changes
 
 - The `SupModel` and `UnSupModel` traits now return
 `Result`s for the `train` and `predict` functions.
@@ -27,17 +93,17 @@ This is another fairly large release. Thank you to everyone who contributed!
 rulinalg changelog for
 details](https://github.com/AtheMathmo/rulinalg/blob/master/CHANGELOG.md#030)).
 
-## Features
+### Features
 
 - Adding RMSProp gradient descent algorithm. #121
 - Adding cross validation. #125
 - Adding a new `Shuffler` transformer. #135
 
-## Bug Fixes
+### Bug Fixes
 
 - None
 
-## Minor Changes
+### Minor Changes
 
 - Adding benchmarks
 - Initiate GMM with sample covariance of data (instead of identity matrix).